+++ src/secrets/mod.rs
@@ -8,5 +8,7 @@ pub mod kv1;
 pub mod kv2;
 #[cfg(feature = "pki")]
 pub mod pki;
+#[cfg(feature = "totp")]
+pub mod totp;
 #[cfg(feature = "transit")]
 pub mod transit;